==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Describes the acquisition process of the datasets used
# Author: Bella MacLean
# Date: 28 March 2024
# Contact: bella.maclean@mail.utoronto.ca
# License: MIT
# Pre-requisites: No
# Any other information needed?: No


#### Download data ####

## Average home matchday attendance data downloaded from: https://www.footballwebpages.co.uk/premier-league/attendances/2022-2023
## Current market value downloaded from: https://www.transfermarkt.us/premier-league/daten/wettbewerb/GB1
## Total payroll downloaded from: https://www.spotrac.com/epl/payroll/
## Total transfer fees downloaded from: https://www.spotrac.com/epl/payroll/
## League standings downloaded from: https://www.transfermarkt.us/premier-league/daten/wettbewerb/GB1
